+++ config/session.ts
@@ -4,16 +4,15 @@ import {
 } from "@innovatedev/fresh-session";
 import { KvDexSessionStorage } from "@innovatedev/fresh-session/kvdex-store";
 import type { State } from "@/utils.ts";
+import { SessionData, User } from "@/kv/models.ts";
 import { db } from "@/kv/db.ts";
 import { rulesToPermissions } from "@/lib/permissions.ts";
 import settings from "@/config/app.ts";
 
-export const sessionConfig: SessionOptions = {
-  store: new KvDexSessionStorage({
-    // deno-lint-ignore no-explicit-any
-    collection: db.sessions as any,
-    // deno-lint-ignore no-explicit-any
-    userCollection: db.users as any,
+export const sessionConfig: SessionOptions<User, SessionData> = {
+  store: new KvDexSessionStorage<SessionData, User>({
+    collection: db.sessions,
+    userCollection: db.users,
     expireAfter: 60 * 60 * 24 * 7, // 1 week
     // userIndex: "email", // Optional secondary index
   }),
@@ -21,11 +20,11 @@ export const sessionConfig: SessionOptions = {
     name: "innokv.sid",
     httpOnly: true,
     secure: settings.env.isProd,
-    sameSite: "Lax",
+    sameSite: "Lax" as const,
     maxAge: 60 * 60 * 24 * 7, // 1 week
   },
   // Enable for Stateless API Token Support (e.g. "Authorization: Bearer <token>")
-  verifyToken: async (tokenSecret) => {
+  verifyToken: async (tokenSecret: string) => {
     // Hash with SHA-256 to find
     const encoder = new TextEncoder();
     const dataBytes = encoder.encode(tokenSecret);
@@ -39,11 +38,11 @@ export const sessionConfig: SessionOptions = {
       tokenHash,
     );
 
-    if (!tokenDoc) return null;
+    if (!tokenDoc) return undefined;
 
     // Check expiration
     if (tokenDoc.value.expiresAt && new Date() > tokenDoc.value.expiresAt) {
-      return null;
+      return undefined;
     }
 
     // Update lastUsedAt
@@ -54,7 +53,7 @@ export const sessionConfig: SessionOptions = {
 
     // Populate user
     const userDoc = await db.users.find(tokenDoc.value.userId);
-    if (!userDoc) return null;
+    if (!userDoc) return undefined;
 
     return {
       id: userDoc.id,